Split desktop ui/ui-text into desktopMain and skikoMain parts
Test: jvmTest, desktopTest
