+++ compose/ui/ui-text/src/desktopMain/kotlin/androidx/compose/ui/text/platform/WeakHashMap.desktop.kt
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2020 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package androidx.compose.ui.text.platform
+
+// Suppressing the diagnostics caused by https://youtrack.jetbrains.com/issue/KT-37316
+@Suppress("ACTUAL_WITHOUT_EXPECT")
+internal actual typealias WeakHashMap<K, V> = java.util.WeakHashMap<K, V>
